requirements.txt
asgiref==3.2.3
Django==3.0.1
django-bootstrap-form==3.4
django-crispy-forms==1.8.1
django-filter==2.2.0
django-widget-tweaks==1.4.5
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
